pyproject.toml
[tool.poetry]
name = "nirscloud-julia"
version = "0.0.7"
description = "Julia's nirscloud utils"
authors = ["Julia Tatz <tatz.j@northeastern.edu>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
numpy = "~1.21"
scipy = "~1.7"
numpy-quaternion = "~2021.7"
xarray = "~0.19"
pymongo = "~3.12"
matplotlib = "~3.4"
pandas = "~1.3"
zarr = "~2.8"
requests = "~2.26"
hdfs = "~2.6"
requests-gssapi = "~1.2"
numba = "~0.53"
fastparquet = "^0.7.1"

[tool.poetry.dev-dependencies]
ipywidgets = ">=7.6"
PySide2 = "^5.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
